+++ code/Online/Local/LocalStatistics.h
@@ -24,7 +24,7 @@ public:
 
 	virtual bool enumerate(std::map< std::wstring, float >& outStats);
 
-	virtual bool set(const std::wstring& statId, float value);
+	virtual bool set(const std::wstring& statId, int32_t value);
 
 private:
 	Ref< sql::IConnection > m_db;
